+++ src/main/java/com/example/gradle/p1/exceptions/PrimeNotFoundException.java
@@ -2,14 +2,14 @@ package com.example.gradle.p1.exceptions;
 
 public class PrimeNotFoundException extends RuntimeException {
 
-    private Long primeToSearchTo;
+    private Integer primeToSearchTo;
 
-    public PrimeNotFoundException(final Long primeToSearchTo) {
+    public PrimeNotFoundException(final Integer primeToSearchTo) {
         super("No Prime numbers found below " + primeToSearchTo);
         this.primeToSearchTo = primeToSearchTo;
     }
 
-    public Long getPrimeToSearchTo() {
+    public Integer getPrimeToSearchTo() {
         return primeToSearchTo;
     }
 }
